fix `$setBlocksType` not rendering Decorator
- if change the Para style of a block with a verse, previously the verse number would disappear (not get re-rendered).
- change `ImmutableVerseNode` to TSX so React re-renders it.
