+++ packages/scribe/src/plugins/ScriptureReferencePlugin.tsx
@@ -13,11 +13,10 @@ import {
   findChapter,
   findNextChapter,
   findThisChapter,
-  findThisVerse,
-  findVerse,
   removeNodeAndAfter,
   removeNodesBeforeNode,
 } from "shared/nodes/scripture/usj/node.utils";
+import { findThisVerse, findVerse } from "shared-react/nodes/scripture/usj/node-react.utils";
 import { BookCode } from "utilities/src/converters/usj/usj.model.ts";
 import {
   getChapterNodeClass,
